requirements.txt
django-heroku
gunicorn
whitenoise
requests